pyproject.toml
[tool.poetry]
name = "google-sheets-llm-analyzer"
version = "1.0.0"
description = "LLM-integrated Google Sheets data processing via API and analysis toolkit"
authors = ["script-logic <dev.scriptlogic@gmail.com>"]
readme = "README.md"
license = "MIT"
homepage = "https://github.com/script-logic/google-sheets-llm-analyzer"
repository = "https://github.com/script-logic/google-sheets-llm-analyzer"
documentation = "https://github.com/script-logic/google-sheets-llm-analyzer"
keywords = ["google-sheets", "llm", "data-analysis", "openai", "automation"]
classifiers = [
    "Topic :: Office/Business",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[tool.poetry.dependencies]
python = ">=3.11"
"google-api-python-client" = ">=2.108.0"
"google-auth" = ">=2.23.0"
"google-auth-httplib2" = ">=0.1.0"
"google-auth-oauthlib" = ">=1.0.0"
openai = ">=1.6.0"
pydantic = ">=2.5.0"
"pydantic-settings" = ">=2.1.0"
rich = ">=13.0.0"

[tool.poetry.group.dev.dependencies]
mypy = ">=1.15.0"
ruff = ">=0.4.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py311"
line-length = 79
format.quote-style = "double"
lint.select = [
    "E",
    "W",
    "F",
    "I",
    "B",
    "C4",
    "UP",
    "RUF",
]
exclude = [
    ".venv/",
    "__pycache__/",
    ".git/",
    "*.pyc",
]

[tool.ruff.lint.isort]
known-first-party = ["google_sheets_llm_analyzer"]

[tool.mypy]
python_version = "3.11"
warn_return_any = false
disable_error_code = ["import-untyped"]